+++ apps/todos/src/app/features/todos/containers/index.ts
@@ -1,1 +1,1 @@
-export * from './todos-container';
+export * from './todos';
